# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Sevu API - Customer management, WhatsApp reminder system, and public profiles
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /reminders/log:
    post:
      operationId: logReminderSent
      tags: [reminders]
      summary: Log that a reminder was sent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogReminderBody"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderLog"
  /reviews/log:
    post:
      operationId: logReviewSent
      tags: [reminders]
      summary: Log that a review request was sent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogReviewBody"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewLog"
  /service-logs:
    get:
      operationId: listServiceLogs
      tags: [services]
      summary: List service logs for a business
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of service logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceLog"
    post:
      operationId: createServiceLog
      tags: [services]
      summary: Create a service log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceLogBody"
      responses:
        "201":
          description: Created service log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceLog"
  /service-logs/{id}:
    put:
      operationId: updateServiceLog
      tags: [services]
      summary: Update a service log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceLogBody"
      responses:
        "200":
          description: Updated service log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceLog"
    delete:
      operationId: deleteServiceLog
      tags: [services]
      summary: Delete a service log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        phone:
          type: string
        address:
          type: string
        reviewLink:
          type: string
        defaultReminderMessage:
          type: string
        defaultReviewMessage:
          type: string
        createdAt:
          type: string
      required: [id, name, category]
...
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        phone:
          type: string
        address:
          type: string
        reviewLink:
          type: string
        defaultReminderMessage:
          type: string
        defaultReviewMessage:
          type: string
...
    LogReminderBody:
      type: object
      properties:
        customerId:
          type: integer
        businessId:
          type: integer
      required: [customerId, businessId]
    LogReviewBody:
      type: object
      properties:
        customerId:
          type: integer
        businessId:
          type: integer
      required: [customerId, businessId]
    ReminderLog:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        businessId:
          type: integer
        sentAt:
          type: string
      required: [id, customerId, businessId, sentAt]
    ReviewLog:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        businessId:
          type: integer
        sentAt:
          type: string
      required: [id, customerId, businessId, sentAt]
...
    ServiceLog:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        service:
          type: string
        amount:
          type: number
        paymentStatus:
          type: string
        serviceDate:
          type: string
        nextVisit:
          type: string
        note:
          type: string
        createdAt:
          type: string
      required: [id, businessId, customerName, service, paymentStatus, serviceDate, createdAt]
    CreateServiceLogBody:
      type: object
      properties:
        businessId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        service:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        paymentStatus:
          type: string
        serviceDate:
          type: string
        paymentDate:
          type: string
        nextVisit:
          type: string
        note:
          type: string
      required: [businessId, customerName, service, paymentStatus, serviceDate]
    UpdateServiceLogBody:
      type: object
      properties:
        service:
          type: string
        amount:
          type: number
        paymentStatus:
          type: string
        serviceDate:
          type: string
        nextVisit:
          type: string
        note:
          type: string
...